Starting wiring up search UX -> search provider
