+++ browser/src/Plugins/Api/Oni.ts
@@ -27,7 +27,6 @@ import { getInstance as getMenuManagerInstance } from "./../../Services/Menu"
 import { getInstance as getNotificationsInstance } from "./../../Services/Notifications"
 import { getInstance as getOverlayInstance } from "./../../Services/Overlay"
 import { recorder } from "./../../Services/Recorder"
-import { getInstance as getSearchInstance } from "./../../Services/Search"
 import { getInstance as getSidebarInstance } from "./../../Services/Sidebar"
 import { getInstance as getSnippetsInstance } from "./../../Services/Snippets"
 import { getInstance as getStatusBarInstance } from "./../../Services/StatusBar"
@@ -125,10 +124,6 @@ export class Oni extends EventEmitter implements OniApi.Plugin.Api {
 
     public get process(): OniApi.Process {
         return Process
-    }
-
-    public get search(): any {
-        return getSearchInstance()
     }
 
     public get sidebar(): any {
