+++ browser/src/index.tsx
@@ -187,9 +187,6 @@ const start = async (args: string[]): Promise<void> => {
 
     diagnostics.start(languageManager)
 
-    const Search = await import("./Services/Search")
-    Search.activate()
-
     Performance.startMeasure("Oni.Start.Activate")
     const api = pluginManager.startApi()
     configuration.activate(api)
